tools/bridge-stub/pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "grok-build-bridge"
version = "0.1.0+stub"
description = "CI-only stub for grok-build-bridge — see tools/bridge-stub/README.md."
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = [
    "pyyaml>=6,<7",
    "rich>=13,<15",
]

[tool.hatch.build.targets.wheel]
packages = ["grok_build_bridge"]
